refactor: network configuration state machine (#956)
