+++ src/stateMachines/wiredNetworkConfiguration.ts
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  **********************************************************************/
 
-import { type IPS, type AMT } from '@open-amt-cloud-toolkit/wsman-messages'
+import { type AMT, type CIM, type IPS } from '@open-amt-cloud-toolkit/wsman-messages'
 import { assign, createMachine, send } from 'xstate'
 import { type HttpHandler } from '../HttpHandler'
 import Logger from '../Logger'
@@ -12,35 +12,39 @@ import { devices } from '../WebSocketListener'
 import { Error } from './error'
 import { Configurator } from '../Configurator'
 import { DbCreatorFactory } from '../factories/DbCreatorFactory'
-import { invokeEnterpriseAssistantCall, invokeWsmanCall } from './common'
+import { invokeWsmanCall } from './common'
 import { UNEXPECTED_PARSE_ERROR } from '../utils/constants'
-import { type EnterpriseAssistantMessage } from '../WSEnterpriseAssistantListener'
+import { getCertFromEnterpriseAssistant, initiateCertRequest, sendEnterpriseAssistantKeyPairResponse } from './enterpriseAssistant'
 
-export interface IEEE8021xConfigContext {
+interface WiredConfigContext {
   amtProfile: AMTConfiguration
   ieee8021xProfile: any
   message: any
   clientId: string
   xmlMessage: any
   statusMessage: string
-  httpHandler: HttpHandler
-  targetAfterError: string
   errorMessage: string
-  retryCount?: number
-  ips?: IPS.Messages
+  generalSettings: AMT.Models.GeneralSettings
+  wiredSettings: any
+  httpHandler: HttpHandler
   amt?: AMT.Messages
+  ips?: IPS.Messages
+  cim?: CIM.Messages
+  retryCount?: number
   eaResponse: any
   addTrustedRootCertResponse: any
   addCertResponse: any
   keyPairHandle?: string
+  targetAfterError: string
+  wirelessSettings: any
 }
 
-interface IEEE8021xConfigEvent {
-  type: 'CONFIGURE_8021X' | 'ONFAILED'
+interface WiredConfigEvent {
+  type: 'WIREDCONFIG' | 'ONFAILED'
   clientId: string
   data?: any
 }
-export class IEEE8021xConfiguration {
+export class WiredConfiguration {
   configurator: Configurator
   logger: Logger
   dbFactory: DbCreatorFactory
@@ -48,7 +52,7 @@ export class IEEE8021xConfiguration {
   error: Error = new Error()
 
   machine =
-    createMachine<IEEE8021xConfigContext, IEEE8021xConfigEvent>({
+    createMachine<WiredConfigContext, WiredConfigEvent>({
       preserveActionOrder: true,
       predictableActionArguments: true,
       context: {
@@ -58,6 +62,9 @@ export class IEEE8021xConfiguration {
         clientId: '',
         xmlMessage: null,
         statusMessage: '',
+        generalSettings: null,
+        wiredSettings: null,
+        wirelessSettings: null,
         errorMessage: '',
         targetAfterError: null,
         ieee8021xProfile: null,
@@ -65,20 +72,38 @@ export class IEEE8021xConfiguration {
         addTrustedRootCertResponse: null,
         addCertResponse: null
       },
-      id: 'IEEE8021x-configuration-machine',
+      id: 'wired-network-configuration-machine',
       initial: 'ACTIVATION',
       states: {
         ACTIVATION: {
           on: {
-            CONFIGURE_8021X: {
-              actions: ['Reset Unauth Count', 'Reset Retry Count'],
-              target: 'CHECK_8021X_PROFILE'
+            WIREDCONFIG: {
+              actions: [assign({ statusMessage: () => '' }), 'Reset Unauth Count', 'Reset Retry Count'],
+              target: 'PUT_ETHERNET_PORT_SETTINGS'
             }
           }
         },
-        CHECK_8021X_PROFILE: {
+        PUT_ETHERNET_PORT_SETTINGS: {
+          invoke: {
+            src: this.putEthernetPortSettings.bind(this),
+            id: 'put-ethernet-port-settings',
+            onDone: {
+              actions: assign({ message: (context, event) => event.data }),
+              target: 'CHECK_ETHERNET_PORT_SETTINGS_PUT_RESPONSE'
+            },
+            onError: {
+              actions: assign({ errorMessage: (context, event) => 'Failed to put to ethernet port settings' }),
+              target: 'FAILED'
+            }
+          }
+        },
+        CHECK_ETHERNET_PORT_SETTINGS_PUT_RESPONSE: {
           always: [
             {
+              cond: 'isNotEthernetSettingsUpdated',
+              actions: assign({ errorMessage: (context, event) => 'Failed to put to ethernet port settings' }),
+              target: 'FAILED'
+            }, {
               cond: 'is8021xProfilesExists',
               target: 'GET_8021X_PROFILE'
             }, {
@@ -91,18 +116,18 @@ export class IEEE8021xConfiguration {
             src: this.get8021xProfile.bind(this),
             id: 'get-8021x-profile',
             onDone: {
-              actions: [assign({ message: (context, event) => event.data }), 'Reset Unauth Count'],
+              actions: assign({ ieee8021xProfile: (context, event) => event.data.Envelope.Body.IPS_IEEE8021xSettings }),
               target: 'ENTERPRISE_ASSISTANT_REQUEST'
             },
             onError: {
-              actions: assign({ message: (context, event) => event.data, targetAfterError: (context, event) => 'GET_8021X_PROFILE' }),
+              actions: assign({ errorMessage: 'Failed to get 8021x wired profile' }),
               target: 'FAILED'
             }
           }
         },
         ENTERPRISE_ASSISTANT_REQUEST: {
           invoke: {
-            src: this.initiateCertRequest.bind(this),
+            src: async (context, event) => await initiateCertRequest(context, event),
             id: 'enterprise-assistant-request',
             onDone: {
               actions: [
@@ -121,15 +146,11 @@ export class IEEE8021xConfiguration {
             src: this.generateKeyPair.bind(this),
             id: 'generate-key-pair',
             onDone: {
-              actions: [
-                assign({ message: (context, event) => event.data })
-              ],
+              actions: assign({ message: (context, event) => event.data }),
               target: 'ENUMERATE_PUBLIC_PRIVATE_KEY_PAIR'
             },
             onError: {
-              actions: assign({
-                errorMessage: 'Failed to generate key pair in 802.1x'
-              }),
+              actions: assign({ errorMessage: 'Failed to generate key pair in 802.1x' }),
               target: 'FAILED'
             }
           }
@@ -139,15 +160,11 @@ export class IEEE8021xConfiguration {
             src: this.enumeratePublicPrivateKeyPair.bind(this),
             id: 'enumerate-public-private-key-pair',
             onDone: {
-              actions: [
-                assign({ message: (context, event) => event.data })
-              ],
+              actions: assign({ message: (context, event) => event.data }),
               target: 'PULL_PUBLIC_PRIVATE_KEY_PAIR'
             },
             onError: {
-              actions: assign({
-                errorMessage: 'Failed to enumerate public private key pair in 802.1x'
-              }),
+              actions: assign({ errorMessage: 'Failed to enumerate public private key pair in 802.1x' }),
               target: 'FAILED'
             }
           }
@@ -172,10 +189,10 @@ export class IEEE8021xConfiguration {
         },
         ENTERPRISE_ASSISTANT_RESPONSE: {
           invoke: {
-            src: this.sendEnterpriseAssistantKeyPairResponse.bind(this),
+            src: async (context, event) => await sendEnterpriseAssistantKeyPairResponse(context, event),
             id: 'enterprise-assistant-response',
             onDone: {
-              actions: [assign({ message: (context, event) => event.data })],
+              actions: assign({ message: (context, event) => event.data }),
               target: 'SIGN_CSR'
             },
             onError: {
@@ -189,7 +206,7 @@ export class IEEE8021xConfiguration {
             src: this.signCSR.bind(this),
             id: 'sign-csr',
             onDone: {
-              actions: [assign({ message: (context, event) => event.data })],
+              actions: assign({ message: (context, event) => event.data }),
               target: 'GET_CERT_FROM_ENTERPRISE_ASSISTANT'
             },
             onError: {
@@ -200,10 +217,10 @@ export class IEEE8021xConfiguration {
         },
         GET_CERT_FROM_ENTERPRISE_ASSISTANT: {
           invoke: {
-            src: this.getCertFromEnterpriseAssistant.bind(this),
+            src: async (context, event) => await getCertFromEnterpriseAssistant(context, event),
             id: 'get-cert-from-enterprise-assistant',
             onDone: {
-              actions: [assign({ message: (context, event) => event.data })],
+              actions: assign({ message: (context, event) => event.data }),
               target: 'ADD_CERTIFICATE'
             },
             onError: {
@@ -217,15 +234,11 @@ export class IEEE8021xConfiguration {
             src: this.addCertificate.bind(this),
             id: 'add-certificate',
             onDone: {
-              actions: [
-                assign({ message: (context, event) => event.data })
-              ],
+              actions: assign({ message: (context, event) => event.data }),
               target: 'ADD_RADIUS_SERVER_ROOT_CERTIFICATE'
             },
             onError: {
-              actions: assign({
-                errorMessage: 'Failed to add certificate in 802.1x'
-              }),
+              actions: assign({ errorMessage: 'Failed to add certificate in 802.1x' }),
               target: 'FAILED'
             }
           }
@@ -235,15 +248,11 @@ export class IEEE8021xConfiguration {
             src: this.addRadiusServerRootCertificate.bind(this),
             id: 'add-radius-server-root-certificate',
             onDone: {
-              actions: [
-                assign({ message: (context, event) => event.data })
-              ],
+              actions: assign({ message: (context, event) => event.data }),
               target: 'PUT_8021X_PROFILE'
             },
             onError: {
-              actions: assign({
-                errorMessage: 'Failed to add radius server root certificate in 802.1x'
-              }),
+              actions: assign({ errorMessage: 'Failed to add radius server root certificate in 802.1x' }),
               target: 'FAILED'
             }
           }
@@ -285,8 +294,7 @@ export class IEEE8021xConfiguration {
               unauthCount: 0,
               message: (context, event) => event.data,
               clientId: (context, event) => context.clientId
-            },
-            onDone: 'GET_8021X_PROFILE'
+            }
           },
           on: {
             ONFAILED: 'FAILED'
@@ -297,37 +305,75 @@ export class IEEE8021xConfiguration {
           type: 'final'
         },
         SUCCESS: {
+          entry: [assign({ statusMessage: 'Wired Network Configured' }), 'Update Configuration Status'],
           type: 'final'
         }
       }
     }, {
       guards: {
+        isNotEthernetSettingsUpdated: (context, event) => {
+          const amtEthernetPortSettings: AMT.Models.EthernetPortSettings = context.message.Envelope.Body.AMT_EthernetPortSettings
+          if (context.amtProfile.dhcpEnabled === amtEthernetPortSettings.DHCPEnabled && !(context.amtProfile.dhcpEnabled) === amtEthernetPortSettings.SharedStaticIp && amtEthernetPortSettings.IpSyncEnabled) {
+            return false
+          }
+          return true
+        },
         is8021xProfilesExists: (context, event) => context.amtProfile.ieee8021xProfileName != null,
-        shouldRetry: (context, event) => context.retryCount < 3 && event.data === UNEXPECTED_PARSE_ERROR
+        shouldRetry: (context, event) => context.retryCount < 3 && event.data instanceof UNEXPECTED_PARSE_ERROR
       },
       actions: {
         'Reset Unauth Count': (context, event) => { devices[context.clientId].unauthCount = 0 },
-        'Update Configuration Status': (context, event) => {
-          const status = devices[context.clientId].status.Network
-          devices[context.clientId].status.Network = status == null ? context.statusMessage : `${status}. ${context.errorMessage}`
-        },
         'Reset Retry Count': assign({ retryCount: (context, event) => 0 }),
-        'Increment Retry Count': assign({ retryCount: (context, event) => context.retryCount + 1 })
+        'Increment Retry Count': assign({ retryCount: (context, event) => context.retryCount + 1 }),
+        'Update Configuration Status': (context, event) => {
+          devices[context.clientId].status.Network = context.errorMessage ?? context.statusMessage
+        }
       }
     })
 
   constructor () {
     this.configurator = new Configurator()
     this.dbFactory = new DbCreatorFactory()
-    this.logger = new Logger('8021x_Configuration_State_Machine')
+    this.logger = new Logger('Network_Configuration_State_Machine')
   }
 
-  async get8021xProfile (context: IEEE8021xConfigContext, event: IEEE8021xConfigEvent): Promise<void> {
+  async putEthernetPortSettings (context): Promise<any> {
+    if (context.amtProfile.dhcpEnabled) {
+      context.wiredSettings.DHCPEnabled = true
+      context.wiredSettings.SharedStaticIp = false
+    } else {
+      context.wiredSettings.DHCPEnabled = false
+      context.wiredSettings.SharedStaticIp = true
+    }
+    context.wiredSettings.IpSyncEnabled = true
+    if (context.wiredSettings.DHCPEnabled || context.wiredSettings.IpSyncEnabled) {
+      // When 'DHCPEnabled' property is set to true the following properties should be removed:
+      // SubnetMask, DefaultGateway, IPAddress, PrimaryDNS, SecondaryDNS.
+      delete context.wiredSettings.SubnetMask
+      delete context.wiredSettings.DefaultGateway
+      delete context.wiredSettings.IPAddress
+      delete context.wiredSettings.PrimaryDNS
+      delete context.wiredSettings.SecondaryDNS
+    } else {
+      // TBD: To set static IP address the values should be read from the REST API
+      // ethernetPortSettings.SubnetMask = "255.255.255.0";
+      // ethernetPortSettings.DefaultGateway = "192.168.1.1";
+      // ethernetPortSettings.IPAddress = "192.168.1.223";
+      // ethernetPortSettings.PrimaryDNS = "192.168.1.1";
+      // ethernetPortSettings.SecondaryDNS = "192.168.1.1";
+    }
+    // this.logger.debug(`Updated Network configuration to set on device :  ${JSON.stringify(context.message, null, '\t')}`)
+    // put request to update ethernet port settings on the device
+    context.xmlMessage = context.amt.EthernetPortSettings.Put(context.wiredSettings)
+    return await invokeWsmanCall(context, 2)
+  }
+
+  async get8021xProfile (context: WiredConfigContext, event: WiredConfigEvent): Promise<void> {
     context.xmlMessage = context.ips.IEEE8021xSettings.Get()
     return await invokeWsmanCall(context, 2)
   }
 
-  async put8021xProfile (context: IEEE8021xConfigContext, event: IEEE8021xConfigEvent): Promise<void> {
+  async put8021xProfile (context: WiredConfigContext, event: WiredConfigEvent): Promise<void> {
     this.logger.info('EA Response :', JSON.stringify(context.eaResponse))
     context.addTrustedRootCertResponse = context.message.Envelope.Body
     context.ieee8021xProfile.Enabled = 2
@@ -339,7 +385,7 @@ export class IEEE8021xConfiguration {
     return await invokeWsmanCall(context, 2)
   }
 
-  async setCertificates (context: IEEE8021xConfigContext, event: IEEE8021xConfigEvent): Promise<void> {
+  async setCertificates (context: WiredConfigContext, event: WiredConfigEvent): Promise<void> {
     const clientCertReference1 = context.addCertResponse?.AddCertificate_OUTPUT?.CreatedCertificate?.ReferenceParameters?.SelectorSet?.Selector?._
     const rootCertReference1 = context.addTrustedRootCertResponse?.AddTrustedRootCertificate_OUTPUT?.CreatedCertificate?.ReferenceParameters?.SelectorSet?.Selector?._
     const rootCertReference = `<a:Address>default</a:Address><a:ReferenceParameters><w:ResourceURI>http://intel.com/wbem/wscim/1/amt-schema/1/AMT_PublicKeyCertificate</w:ResourceURI><w:SelectorSet><w:Selector Name="InstanceID">${rootCertReference1}</w:Selector></w:SelectorSet></a:ReferenceParameters>`
@@ -349,62 +395,23 @@ export class IEEE8021xConfiguration {
     return await invokeWsmanCall(context, 2)
   }
 
-  async initiateCertRequest (context: IEEE8021xConfigContext, event: IEEE8021xConfigEvent): Promise<EnterpriseAssistantMessage> {
-    context.ieee8021xProfile = context.message.Envelope.Body.IPS_IEEE8021xSettings
-    context.message = {
-      action: 'satellite',
-      subaction: '802.1x-ProFile-Request',
-      satelliteFlags: 2,
-      nodeid: context.clientId,
-      domain: '',
-      reqid: '',
-      authProtocol: 0,
-      osname: 'win11',
-      devname: devices[context.clientId].hostname,
-      icon: 1,
-      cert: null,
-      certid: null,
-      ver: ''
-    }
-    return await invokeEnterpriseAssistantCall(context)
-  }
-
-  async generateKeyPair (context: IEEE8021xConfigContext, event: IEEE8021xConfigEvent): Promise<void> {
+  async generateKeyPair (context: WiredConfigContext, event: WiredConfigEvent): Promise<void> {
     context.xmlMessage = context.amt.PublicKeyManagementService.GenerateKeyPair({ KeyAlgorithm: 0, KeyLength: 2048 })
     return await invokeWsmanCall(context)
   }
 
-  async enumeratePublicPrivateKeyPair (context: IEEE8021xConfigContext, event: IEEE8021xConfigEvent): Promise<void> {
+  async enumeratePublicPrivateKeyPair (context: WiredConfigContext, event: WiredConfigEvent): Promise<void> {
     context.keyPairHandle = context.message.Envelope.Body?.GenerateKeyPair_OUTPUT?.KeyPair?.ReferenceParameters?.SelectorSet?.Selector?._
     context.xmlMessage = context.amt.PublicPrivateKeyPair.Enumerate()
     return await invokeWsmanCall(context, 2)
   }
 
-  async pullPublicPrivateKeyPair (context: IEEE8021xConfigContext, event: IEEE8021xConfigEvent): Promise<void> {
+  async pullPublicPrivateKeyPair (context: WiredConfigContext, event: WiredConfigEvent): Promise<void> {
     context.xmlMessage = context.amt.PublicPrivateKeyPair.Pull(context.message.Envelope.Body?.EnumerateResponse?.EnumerationContext)
     return await invokeWsmanCall(context)
   }
 
-  async getCertFromEnterpriseAssistant (context: IEEE8021xConfigContext, event: IEEE8021xConfigEvent): Promise<EnterpriseAssistantMessage> {
-    const signedCSR = context.message.Envelope?.Body?.GeneratePKCS10RequestEx_OUTPUT?.SignedCertificateRequest
-    context.message = {
-      action: 'satellite',
-      subaction: '802.1x-CSR-Response',
-      satelliteFlags: 2,
-      nodeid: context.clientId,
-      domain: '',
-      reqid: '',
-      authProtocol: 0,
-      osname: 'win11',
-      devname: devices[context.clientId].hostname,
-      icon: 1,
-      signedcsr: signedCSR,
-      ver: ''
-    }
-    return await invokeEnterpriseAssistantCall(context)
-  }
-
-  async signCSR (context: IEEE8021xConfigContext, event: IEEE8021xConfigEvent): Promise<void> {
+  async signCSR (context: WiredConfigContext, event: WiredConfigEvent): Promise<void> {
     context.xmlMessage = context.amt.PublicKeyManagementService.GeneratePKCS10RequestEx({
       KeyPair: '<a:Address>http://schemas.xmlsoap.org/ws/2004/08/addressing/role/anonymous</a:Address><a:ReferenceParameters><w:ResourceURI>http://intel.com/wbem/wscim/1/amt-schema/1/AMT_PublicPrivateKeyPair</w:ResourceURI><w:SelectorSet><w:Selector Name="InstanceID">' + (context.message.response.keyInstanceId as string) + '</w:Selector></w:SelectorSet></a:ReferenceParameters>',
       SigningAlgorithm: 1,
@@ -413,43 +420,14 @@ export class IEEE8021xConfiguration {
     return await invokeWsmanCall(context)
   }
 
-  async sendEnterpriseAssistantKeyPairResponse (context: IEEE8021xConfigContext, event: IEEE8021xConfigEvent): Promise<EnterpriseAssistantMessage> {
-    const clientObj = devices[context.clientId]
-    const potentialArray = context.message.Envelope.Body.PullResponse.Items.AMT_PublicPrivateKeyPair
-    if (Array.isArray(potentialArray)) {
-      clientObj.tls.PublicPrivateKeyPair = potentialArray
-    } else {
-      clientObj.tls.PublicPrivateKeyPair = [potentialArray]
-    }
-    const PublicPrivateKeyPair = clientObj.tls.PublicPrivateKeyPair.filter(x => x.InstanceID === context.keyPairHandle)[0]
-    const DERKey = PublicPrivateKeyPair?.DERKey
-
-    context.message = {
-      action: 'satellite',
-      subaction: '802.1x-KeyPair-Response',
-      satelliteFlags: 2,
-      nodeid: context.clientId,
-      domain: '',
-      reqid: '',
-      devname: devices[context.clientId].hostname,
-      authProtocol: 0,
-      osname: 'win11',
-      icon: 1,
-      DERKey,
-      keyInstanceId: PublicPrivateKeyPair?.InstanceID,
-      ver: ''
-    }
-    return await invokeEnterpriseAssistantCall(context)
-  }
-
-  async addCertificate (context: IEEE8021xConfigContext, event: IEEE8021xConfigEvent): Promise<void> {
+  async addCertificate (context: WiredConfigContext, event: WiredConfigEvent): Promise<void> {
     context.eaResponse = event.data.response
     const cert = event.data.response.certificate
     context.xmlMessage = context.amt.PublicKeyManagementService.AddCertificate({ CertificateBlob: cert })
     return await invokeWsmanCall(context)
   }
 
-  async addRadiusServerRootCertificate (context: IEEE8021xConfigContext, event: IEEE8021xConfigEvent): Promise<void> {
+  async addRadiusServerRootCertificate (context: WiredConfigContext, event: WiredConfigEvent): Promise<void> {
     // To Do: Needs to replace the logic with how we will pull the radius server root certificate
     context.addCertResponse = context.message.Envelope.Body
     const cert = context.eaResponse.rootcert
